requirements.txt
vk-api==11.9.9
openai==1.12.0
python-dotenv==1.0.0
pandas==2.2.0
SQLAlchemy==2.0.27
openpyxl==3.1.2
python-telegram-bot==20.8
numpy==1.26.4
scikit-learn==1.4.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 